+++ doc/changelog.rst
@@ -1,7 +1,9 @@
 Changelog
 =========
-Changes in Version 4.15.0 (XXXX/XX/XX)
+
+Changes in Version 4.15.0 (2025/09/10)
 --------------------------------------
+
 PyMongo 4.15 brings a number of changes including:
 
 - Added :class:`~pymongo.encryption_options.TextOpts`,
@@ -13,6 +15,7 @@ PyMongo 4.15 brings a number of changes including:
   ``pymongocrypt>=1.16`` is required for text query support.
 - Added :class:`bson.decimal128.DecimalEncoder` and :class:`bson.decimal128.DecimalDecoder`
   to support encoding and decoding of BSON Decimal128 values to decimal.Decimal values using the TypeRegistry API.
+- Added support for Windows ``arm64`` wheels.
 
 Changes in Version 4.14.1 (2025/08/19)
 --------------------------------------